+++ common/src/main/java/net/mehvahdjukaar/vista/client/VistaClientPlatStuff.java
@@ -1,0 +1,26 @@
+package net.mehvahdjukaar.vista.client;
+
+import com.mojang.blaze3d.vertex.PoseStack;
+import net.mehvahdjukaar.candlelight.api.PlatformImpl;
+import net.mehvahdjukaar.vista.client.web.PcmSource;
+import net.mehvahdjukaar.vista.client.web.TvSpeakerSound;
+import net.minecraft.client.Camera;
+import net.minecraft.client.Minecraft;
+import net.minecraft.world.phys.Vec3;
+import org.jetbrains.annotations.Contract;
+import org.joml.Matrix4f;
+
+public class VistaClientPlatStuff {
+
+    @Contract
+    @PlatformImpl
+    public static void dispatchRenderStageAfterLevel(Minecraft mc, PoseStack poseStack, Camera camera, Matrix4f modelViewMatrix, Matrix4f projMatrix) {
+        throw new AssertionError();
+    }
+
+    @Contract
+    @PlatformImpl
+    public static TvSpeakerSound createTvSpeakerSound(PcmSource source, Vec3 pos, double startSeconds) {
+        throw new AssertionError();
+    }
+}