+++ changelog.md
@@ -1,1 +1,1 @@
-- more audio fixes 2
+- fixed a crash on servers